backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
